Split blocks clipboard stuff to BlocksClipboard component
